backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
fuzzywuzzy>=0.18.0
python-levenshtein>=0.23.0
groq>=0.4.0
pymongo>=4.6.0
motor>=3.3.0
